score 수식 변경
